Rename 'main' package to 'service' and configure gradle for JavaFX
