+++ src/main/java/evan/command/DeadlineCommand.java
@@ -1,8 +1,8 @@
 package evan.command;
 
-import evan.main.Storage;
-import evan.main.TaskList;
-import evan.main.Ui;
+import evan.service.Storage;
+import evan.service.TaskList;
+import evan.service.Ui;
 import evan.task.Deadline;
 
 /**
