+++ src/main/java/evan/gui/Launcher.java
@@ -1,0 +1,11 @@
+package evan.gui;
+
+import javafx.application.Application;
+
+/**
+ * A launcher class to work around classpath issues.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+    }
+}
