+++ src/main/java/evan/service/UserInputParser.java
@@ -1,4 +1,4 @@
-package evan.main;
+package evan.service;
 
 import static java.lang.Integer.parseInt;
